server/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
Pillow==10.2.0
pytesseract==0.3.10
openai==1.35.7
python-dotenv==1.0.1
